fix(api): cap narrative length on records PUT and share the validator
Addresses review on #8: the abuse vector was only half-closed. POST got
the 5000-char cap but PUT /api/records/[id] wrote story/actions/results/
feelings/reaction with no limit, so "create small, then edit to 5MB"
still bloated the row and search_vector.

- Extract MAX_NARRATIVE_LENGTH + validateRecordInput into
  src/lib/records-validation.ts (single source of truth; also a clean,
  pure home for Phase 3 unit tests).
- PUT now rejects oversized narrative fields with a 400 fieldTooLong,
  reusing the same cap.
